feat(common): move common files to internal
