+++ client/client.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/gorilla/websocket"
-	"github.com/omarahm3/squirrel/common"
+	"github.com/omarahm3/squirrel/internal/pkg/common"
 	"go.uber.org/zap"
 )
