+++ client/options.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/omarahm3/squirrel/common"
+	"github.com/omarahm3/squirrel/internal/pkg/common"
 	"go.uber.org/zap/zapcore"
 )
