+++ internal/pkg/common/message.go
